pyproject.toml
[tool.poetry]
name = "nexusbridge"
version = "0.1.2"
description = "A unified trading interface for major cryptocurrency exchanges, supporting REST APIs and WebSocket connections for Binance, OKX, Bybit and more."
authors = ["River-Shi <nachuan.shi.quant@gmail.com>"]
license = "MIT LICENSE"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
aiohttp = "^3.11.11"
pathlib = "^1.0.1"
loguru = "^0.7.3"
orjson = "^3.10.12"
dynaconf = "^3.2.6"
certifi = "^2024.12.14"
picows = "^1.7.2"
aiolimiter = "^1.2.1"
pycryptodome = "^3.21.0"
asynciolimiter = "^1.1.1"
httpx = "^0.28.1"
anchorpy = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
